refactor(micro): move D1 schema, migrations, and drizzle config into micro package
- Move db/schema.ts, db/migrations/, and drizzle.config.ts into packages/micro/
- Add loader-db.ts with a D1 HTTP adapter so drizzle can be used at build time
- Rewrite micro-loader to import getMicroPosts from @just-be/micro/loader-db
  instead of hitting the Cloudflare API directly with raw SQL
- Add exports field to micro package.json for the loader-db entrypoint
- Add mise tasks for micro:db:generate and micro:db:migrate
- Remove drizzle-kit, drizzle-orm, and db:* scripts from root package.json
