+++ package.json
@@ -25,9 +25,7 @@
     "test": "bun test",
     "gen-codes": "bun scripts/gen-codes.ts",
     "lint": "oxlint",
-    "lint:fix": "oxlint --fix",
-    "db:generate": "drizzle-kit generate",
-    "db:migrate": "drizzle-kit migrate"
+    "lint:fix": "oxlint --fix"
   },
   "dependencies": {
     "@ascorbic/feed-loader": "^2.0.1",
@@ -42,8 +40,6 @@
     "@webtui/css": "^0.1.5",
     "astro": "6.1.4",
     "beautiful-mermaid": "^1.0.0",
-    "drizzle-kit": "^0.31.9",
-    "drizzle-orm": "^0.45.1",
     "js-yaml": "^4.1.1",
     "marked": "^17.0.1",
     "mdast-util-mdx-jsx": "^3.2.0",
