+++ packages/micro/src/db.ts
@@ -1,29 +1,13 @@
 import { drizzle } from "drizzle-orm/d1";
-import { integer, sqliteTable, text } from "drizzle-orm/sqlite-core";
-import { sql } from "drizzle-orm";
+import { getPlatformProxy } from "wrangler";
+import { microPosts } from "./schema";
 
-// Re-define the schema locally for the CLI
-export const microPosts = sqliteTable("micro_posts", {
-  id: integer("id").primaryKey({ autoIncrement: true }),
-  content: text("content", { length: 280 }).notNull(),
-  createdAt: integer("created_at", { mode: "timestamp" })
-    .notNull()
-    .default(sql`(unixepoch())`),
-  updatedAt: integer("updated_at", { mode: "timestamp" })
-    .notNull()
-    .default(sql`(unixepoch())`)
-    .$onUpdate(() => new Date()),
-  syndicatedTo: text("syndicated_to", { mode: "json" })
-    .$type<string[]>()
-    .default(sql`'[]'`),
-});
+export { microPosts };
 
 export async function getD1Database() {
   // Use wrangler to get local D1 access
   // Find the project root (where wrangler.toml is)
   const projectRoot = import.meta.dir + "/../../..";
-
-  const { getPlatformProxy } = await import("wrangler");
   const { env, dispose } = await getPlatformProxy<{ DB: D1Database }>({
     configPath: `${projectRoot}/wrangler.toml`,
     persist: { path: `${projectRoot}/.wrangler/state/v3` },
