+++ wrangler.toml
@@ -13,4 +13,4 @@ binding = "ASSETS"
 binding = "DB"
 database_name = "micro-blog"
 database_id = "cc5c17ec-ad2f-41b5-8d5c-cf542604c43b"
-migrations_dir = "db/migrations"
+migrations_dir = "packages/micro/db/migrations"
